actualizado a video 615

+++ src/controllers/ProjectController.ts
@@ -8,11 +8,16 @@ export class ProjectController {
     //-ycon await Project.create(req.body) en el try
 
     const project = new Project(req.body);
+
+    //Asigna un manager
+    project.manager = req.user.id;
+
     try {
       await project.save();
-      res.send("Proyecto creado correctamente");
+      return res.send("Proyecto creado correctamente");
     } catch (error) {
       console.log(error);
+      return res.status(500).send("Hubo un error al crear el proyecto");
     }
 
     res.send("Creando un proyecto!");
@@ -20,10 +25,16 @@ export class ProjectController {
 
   static getAllProjects = async (req: Request, res: Response) => {
     try {
-      const projects = await Project.find({});
-      res.json(projects);
+      const projects = await Project.find({
+        $or: [
+          { manager: { $in: req.user.id } },
+          { team: { $in: req.user.id } },
+        ],
+      });
+      return res.json(projects);
     } catch (error) {
       console.log(error);
+      return res.status(500).send("Hubo un error al crear el proyecto");
     }
   };
 
@@ -36,9 +47,19 @@ export class ProjectController {
         const error = new Error("Proyecto no encontrado");
         return res.status(404).json({ error: error.message });
       }
-      res.json(project);
+
+      if (
+        project.manager.toString() !== req.user.id.toString() &&
+        !project.team.includes(req.user.id)
+      ) {
+        const error = new Error("Acción no válida");
+        return res.status(404).json({ error: error.message });
+      }
+
+      return res.json(project);
     } catch (error) {
       console.log(error);
+      return res.status(500).send("Hubo un error al crear el proyecto");
     }
   };
 
@@ -50,13 +71,20 @@ export class ProjectController {
         const error = new Error("Proyecto no encontrado");
         return res.status(404).json({ error: error.message });
       }
+
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error("Solo el Manager puede actualizar un proyecto");
+        return res.status(404).json({ error: error.message });
+      }
+
       project.clientName = req.body.clientName;
       project.projectName = req.body.projectName;
       project.description = req.body.description;
       await project.save();
-      res.send("Proyecto Actualizado");
+      return res.send("Proyecto Actualizado");
     } catch (error) {
       console.log(error);
+      return res.status(500).send("Hubo un error al crear el proyecto");
     }
   };
 
@@ -69,10 +97,17 @@ export class ProjectController {
         const error = new Error("Proyecto no encontrado");
         return res.status(404).json({ error: error.message });
       }
+
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error("Solo el Manager puede eliminar un proyecto");
+        return res.status(404).json({ error: error.message });
+      }
+
       await project.deleteOne();
-      res.send("Proyecto Eliminado");
+      return res.send("Proyecto Eliminado");
     } catch (error) {
       console.log(error);
+      return res.status(500).send("Hubo un error al crear el proyecto");
     }
   };
 }
